submission/requirements.txt
matplotlib==3.8.4
numpy==1.24.4
Pillow==11.0.0
